Renamed a couple of files and remove scope_settings from self as it's only used once at initialization.
