+++ neurodecode/stream_viewer/_ui_scope_controller.py
