+++ neurodecode/stream_viewer/settings/scope_settings_eeg.ini
